[add]CMD_GF_Private_Room_Info增加了当前人数和最大人数 [modify]微信分享修改了文字 [fix]房间信息改变 [fix]微信唤起程序修改 [fix]声音按键的问题

+++ congcongclient/Classes/ClientHN_THJ/Game/SG/Calculate/SGCalculate.cpp
@@ -55,7 +55,7 @@ void SGCalculate::hide()
 void SGCalculate::update(CMD_GF_Private_End_Info* pInfo)
 {
 	int nodeIndex = 0;
-	for (int index = 0; index < SG_GAME_PLAYER; ++index) {
+	for (size_t index = 0; index < SG_GAME_PLAYER; ++index) {
 		SGPlayer* player = SGGameScence::Instance().getPlayerByChairID(index);
 		if (player->isValid()) {
 			auto playerNode = WidgetFun::getChildWidget(this, utility::toString("SGCalculate_", nodeIndex++));
